324.2. Deleting Properties with Omit(Section 24: REST-Based React Apps)
All right so while that goes I'll flip back over to my stream reducer at the top.

I'm going to import underscore from Lotus and now inside my reducer I'm going to add on a new case of

delete stream and so in this case I'm going to return underscore.

Omit my entire state object and I'm going to tell it what key I want to drop off that object in this

case it's going to be action dot payload.

Now again we do not have to add in the ID here.

No ID.

The reason for that is what we just discussed two seconds ago when we dispatch an action of type.

Delete stream.

The payload is the ID itself.

So we do not have to reference a dot ID property.

OK so that's pretty much it right there.

Now the nice thing about omit is that omit is not going to change the original state object.

Instead it creates a new object with all the properties from states without whatever we passed in as

the action not payload.

So there's no need in this case to create a new object on spread in state and do all that stuff that

you're used to.

I guess that's how we're going to handle deleting a stream.

Now let's save this.

We'll take another quick pause when we come back the next section.

We're going to handle our last case which is to take a list of records or an array of records and merge

them all into our state object in one go so quick break and I'll see you in just a minute.