327.2. Rendering All Streams(Section 24: REST-Based React Apps)
So now we can flip back over to our component.

I'm going to delete that console log and above my render method.

I'm going to define a new method called you guessed it render list.

So the goal of this method is to take that list of streams and just render them out as a list on the

screen.

So inside of your I'm going to return the result of this stop Propp start streams map.

So for each stream we're going to return some amount of GSX I'm going to make a multi-line GSX block

you're going to place a div with the class name of item and will give it a key of stream Dom.

ID

and then inside there I'm going to place an icon with a class name of large middle aligned icon camera

and claps my sidebar so you can see the whole line.

Now that's just going to be a nice icon camera looking thing just a little bit of decoration.

So I'm going to actually do anything for us or then place a div inside there with a class name of content

and inside that div right now I'm just going to print out the stream title and then after that when

we get that closing curly brace I'm going to placing another div with a class name of description.

And for the description we'll print out the streams description property.

Now let's make sure that we eventually call this thing from the render method.

So don't say my render method I'm going to turn this again into a JSM multi-line GSX block and I'll

give this thing a div How about an H to header.

Let's say something like simply streams and then I'm going to place a div with a class name of UI seld

list.

And inside that thing I'm going to call this dot render list

OK.

So a whole bunch of typing here but really nothing new.

We've gone over Lisp building quite a bit inside this course.

I going to save this now and then I'll flip back over to my browser and I'll see my list of streams

appear.

Awesome.

So the icon on there is kind of a nice little touch I guess.

So this definitely looks pretty close to our mockup over here.

Not bad at all.

Now remember we do want to make sure that these streams.

If I was the creator of a stream I want to see a delete and edit button on there as well.

In addition the header right here for the actual stream title should be clickable so that if I click

on that right there I should go to the stream detail page right here at the Sjostrom component.

So you still have a little bit of work to do on this component.

So let's take a quick pause right here and I'll catch you in the next video.