329.2. Conditionally Showing Edit and Delete(Section 24: REST-Based React Apps)
Let's add on a little bit of styling now just to make sure that this thing looks a little better than

it currently does.

So I'm going to flip back over to my component.

I'm going to delete the div inside of here and I'm going to put in a multi-line GSX lock instead.

So I'm going to place a div with the class name of right floated content and then inside of there we're

going to eventually turn those edit and delete buttons into links that are going to take the user to

some other page.

Remember that was the entire idea here if you clicked on delete or edit it would take you to either

the edit stream form or the delete stream page.

But for right now we'll just place in some buttons just to have something there and once we actually

put those pages together then we can actually hook up some links or functionality to these buttons some

place a button with the class name of UI button primary

closing button tag.

And I'll give this first button about it and I'll do a class name UI button.

Negative negative is going to make the I did it again negative is going to make the button appear red.

So that's going to be the delete button.

And we probably want that wants to be read just so the user understands hey this is kind of like a dangerous

action that's going to delete some resource.

All right.

Now the other thing that we're going to do here this seem a little bit weird but basically to get semantic

UI to style everything correctly we actually have to move this div right here with the two buttons to

the very start of the div with the class name of item.

In other words we have to call the method not down here at the bottom.

We have to call it before the eye tag right there.

In order for semantic UI to style everything appropriately.

So now I'll say this again it looks like I got my code here a little bit reformatted but it's still

the same it was before.

And then if I flip back over to my application.

Yeah there we go.

So I nicely see edit and delete on the right hand side of the string that I created.

So again if we click on these buttons right now nothing happens.

So we're going to eventually have to wire some functionality to them.

Now the other thing I want to check your really quickly is see what happens when we sign out of the

application if we sign out.

Then my current user ID is not going to be equal to this stream's user ID and so I should not see the

buttons anymore.

Once I sign out of the application so I click on the sign out and as a matter of fact those two buttons

go away.

And now if I sign in as the same user again the buttons come back because I'm now logged in.

And again my user ID is equal to the user ID inside of that third stream.

All right so this looks pretty great.

Now there's one less thing we need to take care of we need to make sure that at the bottom of this page

whenever user is logged in we show the button to allow a user to create a stream.

So let's take care of that really quickly in the next video.