330.1. Linking to Stream Creation(Section 24: REST-Based React Apps)
In this video we're going to do one last quick thing to our streamlets component we need to make sure

that if the user is logged in we show a button at the very bottom that says something like create Stream.

And so if a user clicks on this right here they should be taken to a page where they can create a new

stream.

We don't really have a lot of navigation between our different screens right now so this would be pretty

good to put together just to make sure that we have the ability to easily tie together the different

screens of our application.

All right.

So I to flip back over to my component.

And how about right above the render method.

I'm going to define another helper method.

And I'm going to call this thing something like.

How about render create.

So the idea here is that this is going to render a create button of sorts.

Now remember we only want to show this if the user is signed in.

We already have some information about the current seinen status of the user inside the prop's object

of our component in the form of current user ID right here whenever the user is not signed in.

Current User ID is set to No.

And that indicates that the user is not signed into our application now even though we've kind of got

some information about the users current authentication status inside of our application.